update browser test
