+++ app/static/tests/e2e/sessions.etest.ts
@@ -2,18 +2,12 @@ import {
     expect, test, chromium, Page
 } from "@playwright/test";
 import * as os from "os";
+import {writeCode} from "./utils";
 
-/* eslint-disable no-irregular-whitespace */
-const placeholderCode = `# Code for rehydration!
-initial(S) <- N - I_0`;
+const appUrl = "/apps/day1";
+const saveSessionTimeout = 3000;
 
 test.describe("Sessions tests", () => {
-    const loadPageWithWait = async (page: Page) => {
-        await page.goto("/apps/day1");
-        // We need a short wait here to give the browser a chance to save a session id.
-        await page.waitForTimeout(2000);
-    };
-
     test("can navigate to Sessions page from navbar, and load a session", async () => {
         // We need to use a browser with persistent context instead of the default incognito browser so that
         // we can use the session ids in local storage
@@ -21,11 +15,15 @@ test.describe("Sessions tests", () => {
         const browser = await chromium.launchPersistentContext(userDataDir);
         const page = await browser.newPage();
 
-        await loadPageWithWait(page);
+        await page.goto(appUrl);
+        // change code in this session, which we will later reload and check that we can see the code changes
+        await writeCode(page, "## NEW CODE FOR LOADED SESSION");
+        // give the page a chance to save the session to back end
+        await page.waitForTimeout(saveSessionTimeout);
 
-        // We need to load the page twice so we have a current session plus an older session so we can test
-        // rehydrating the older one
-        await loadPageWithWait(page);
+        // Reload the page to get fresh session, and give it time to save new session Id
+        await page.goto(appUrl);
+        await page.waitForTimeout(saveSessionTimeout);
 
         // Get storage state in order to test that something has been written to it - but reading it here also seems
         // to force Playwright to wait long enough for storage values to be available in the next page, so the session
@@ -53,8 +51,8 @@ test.describe("Sessions tests", () => {
         // wait until there is some text in the code editor
         await page.waitForFunction((selector) => !!document.querySelector(selector)?.textContent, editorSelector);
 
-        // Check placeholder code
-        expect(await page.innerText(editorSelector)).toBe(placeholderCode);
+        // Check edited code
+        await expect(await page.innerText(editorSelector)).toBe("## NEW CODE FOR LOADED SESSION");
 
         await browser.close();
     });
