+++ app/static/tests/e2e/utils.ts
@@ -8,3 +8,10 @@ export const uploadCSVData = async (page: Page, data: string) => {
     };
     await page.setInputFiles("#fitDataUpload", file);
 };
+
+export const writeCode = async (page: Page, code: string) => {
+    await page.press(".monaco-editor textarea", "Control+A");
+    await page.press(".monaco-editor textarea", "Delete");
+    await page.fill(".monaco-editor textarea", "");
+    await page.fill(".monaco-editor textarea", code);
+};
